requirements.txt
wheel
addict==2.4.0
PyYAML==6.0.1
aiohttp==3.9.1
SQLAlchemy==2.0.23
alembic==1.13.0
PyMySQL[rsa]==1.1.0
pydantic==2.5.3
pycryptodome==3.19.0
pandas==2.1.4
